Refactor SignTransaction and Checkout success screens
To enable correct routing and to move transaction creation and sending logic into a shared TransactionSender component.
